refactor: split popover and floating ui rebuild into primitive
