+++ packages/components/src/Filter/FilterMultiSelect/subcomponents/MenuPopup/index.ts
@@ -1,1 +1,2 @@
 export * from './MenuPopup'
+export * from './ResponsiveMenuPopup'
